# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "流程服务，提供流程编排、执行、调度功能",
        "title": "流程服务 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/swagger/flow-service",
    "paths": {
...
        "/nodes": {
            "get": {
                "description": "获取系统中注册的所有节点的完整信息，包括元数据和配置模式",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nodes"
                ],
                "summary": "获取所有节点完整信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.NodesListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIResponse"
                        }
                    }
                }
            }
        },
        "/nodes/{id}/dynamic-data": {
            "post": {
                "description": "获取指定节点的动态配置数据，如数据库表名、列名等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nodes"
                ],
                "summary": "获取节点动态配置数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "方法名",
                        "name": "method",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "方法参数",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/nodes.DynamicDataResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIResponse"
                        }
                    }
                }
            }
        },
        "/nodes/{id}/validate": {
            "post": {
                "description": "验证指定节点的配置是否正确",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nodes"
                ],
                "summary": "验证节点配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "节点配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.ValidationResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "操作成功"
                },
                "status": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "controllers.NodeFullInfo": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "description": "基础信息",
                    "type": "string"
                },
                "metadata": {
                    "description": "完整元数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/nodes.NodeMetadata"
                        }
                    ]
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "controllers.NodesListResponse": {
            "type": "object",
            "properties": {
                "categories_info": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/nodes.CategoryInfo"
                    }
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.NodeFullInfo"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/controllers.PaginationInfo"
                },
                "types_info": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/nodes.TypeInfo"
                    }
                }
            }
        },
...
        "controllers.PaginationInfo": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.ValidationResult": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": true
                },
                "error": {
                    "type": "string"
                },
                "node_id": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
        "nodes.CategoryInfo": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "node_count": {
                    "type": "integer"
                }
            }
        },
        "nodes.ConfigField": {
            "type": "object",
            "properties": {
                "data_source": {
                    "description": "动态数据源配置（新增）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/nodes.DataSourceConfig"
                        }
                    ]
                },
                "default": {},
                "description": {
                    "type": "string"
                },
                "enum": {
                    "type": "array",
                    "items": {}
                },
                "items": {
                    "$ref": "#/definitions/nodes.ConfigField"
                },
                "name": {
                    "description": "字段名称（新增）",
                    "type": "string"
                },
                "placeholder": {
                    "type": "string"
                },
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/nodes.ConfigField"
                    }
                },
                "required": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "widget": {
                    "description": "基础UI属性",
                    "type": "string"
                }
            }
        },
        "nodes.ConfigSchema": {
            "type": "object",
            "properties": {
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/nodes.ConfigField"
                    }
                },
                "required": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "nodes.DataSourceConfig": {
            "type": "object",
            "properties": {
                "cache_time": {
                    "description": "缓存时间(秒)，0表示不缓存",
                    "type": "integer"
                },
                "dependencies": {
                    "description": "依赖的其他配置字段",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fallback": {
                    "description": "获取失败时的fallback值"
                },
                "method": {
                    "description": "方法名，如\"GetTableNames\"",
                    "type": "string"
                },
                "parameters": {
                    "description": "方法参数配置",
                    "type": "object",
                    "additionalProperties": true
                },
                "type": {
                    "description": "\"method\", \"api\", \"static\"",
                    "type": "string"
                }
            }
        },
        "nodes.DynamicDataResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "nodes.NodeMetadata": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "创建信息",
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "config_schema": {
                    "description": "配置定义",
                    "allOf": [
                        {
                            "$ref": "#/definitions/nodes.ConfigSchema"
                        }
                    ]
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "description": "基础信息",
                    "type": "string"
                },
                "input_ports": {
                    "description": "输入输出定义",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/nodes.PortDefinition"
                    }
                },
                "name": {
                    "type": "string"
                },
                "output_ports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/nodes.PortDefinition"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "nodes.PortDefinition": {
            "type": "object",
            "properties": {
                "data_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "multiple": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                }
            }
        },
        "nodes.TypeInfo": {
            "type": "object",
            "properties": {
                "node_count": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}